+++ apps/E2E/src/index.consts.ts
@@ -9,6 +9,7 @@ export * from './ColorTokens/consts';
 export * from './common/consts';
 export * from './ContextualMenu/consts';
 export * from './CornerRadiusTokens/consts';
+export * from './Divider/consts';
 export * from './FocusTrapZone/consts';
 export * from './FocusZone/consts';
 export * from './IconLegacy/consts';