+++ packages/components/Menu/src/context/menuContext.ts
@@ -17,6 +17,7 @@ export const MenuContext = React.createContext<MenuContextValue>({
   checked: [],
   defaultChecked: [],
   hasCheckmarks: false,
+  hasTooltips: false,
   isSubmenu: false,
   open: false,
   onCheckedChange: () => false,