+++ packages/components/Menu/src/context/menuListContext.ts
@@ -7,12 +7,14 @@ import type { MenuListState } from '../MenuList/MenuList.types';
  */
 export type MenuListContextValue = Omit<MenuListState, 'props'> & {
   hasCheckmarks: boolean;
+  hasTooltips: boolean;
 };
 
 export const MenuListContext = React.createContext<MenuListContextValue>({
   isCheckedControlled: false,
   checked: {},
   hasCheckmarks: false,
+  hasTooltips: false,
   onCheckedChange: () => false,
   onArrowClose: () => false,
   addRadioItem: () => false,